+++ nppm/index.py
@@ -149,11 +149,15 @@ def version():
     help='Specify explicitly which registry to look for Node.py packages '
       'with. If not specified, all registries will be checked unless they '
       'specify the `default=false` option.')
+@click.option('-f', '--force', is_flag=True,
+    help='Overwrite existing installed packages if it becomes necessary (eg. '
+      'when a package link is broken, the package can no longer be found but '
+      'its install directory still exists).')
 @exit_with_return
 def install(packages, develop, python, develop_python, upgrade, global_,
             ignore_installed, packagedir, root, recursive, info, dev,
             pip_separate_process, pip_use_target_option, save, save_dev,
-            save_ext, verbose, registry):
+            save_ext, verbose, registry, force):
   """
   Installs one or more Node.Py or Pip packages.
   """
@@ -187,6 +191,7 @@ def install(packages, develop, python, develop_python, upgrade, global_,
   installer = get_installer(global_, root, upgrade, pip_separate_process,
       pip_use_target_option, recursive, verbose, registry)
   installer.ignore_installed = ignore_installed
+  installer.force = force
   if info:
     for key in sorted(installer.dirs):
       print('{}: {}'.format(key, installer.dirs[key]))